refactor: move courses to module
