+++ app/src/app/modules/courses/directives/highlight-courses.directive.ts
