+++ app/src/app/modules/courses/pipes/duration.pipe.ts
