+++ app/src/app/modules/courses/types/course.interface.ts
