+++ app/src/app/modules/courses/types/user.interface.ts
