+++ app/src/app/search/search.component.html
@@ -7,13 +7,13 @@
     >
       <span class="p-input-icon-left">
         <i class="pi pi-search"></i>
-        <input
+        <!-- <input
           type="text"
           pInputText
           style="width: 300px; height: 50px"
           [(ngModel)]="enteredSearchValue"
           (input)="onSearchTextChanged()"
-        />
+        /> -->
       </span>
 
       <span class="pl-4"></span>
