+++ docs/Module03.md
@@ -1,0 +1,22 @@
+# [Diasoft] Быстрый старт Angular
+
+<br/>
+
+### Module 03
+
+<!-- ```
+$ ng generate component modules/courses/courses --dry-run
+``` -->
+
+<br/>
+
+```
+$ ng generate module modules/core --dry-run
+$ ng generate module modules/courses --dry-run
+```
+
+<br/>
+
+```
+$ ng generate service services/courses --skip-tests=true --dry-run
+```
